fix(v2.3): resolve fourth-cycle audit findings
Cycle 4 caught 1 P0 + 2 MEDIUM + 4 LOW + 2 NIT. All fixed except
two intentional non-changes (documented at the bottom).

CRITICAL — bulk_import_create_lease was broken in production:

The cycle-2 H9 fix added `perform public.assert_can_create_lease(
p_primary_tenant_id, p_unit_id)` but the helper's signature is
`(p_unit_id, p_primary_tenant_id)` — args swapped. Even with the
correct order the helper would still fail every bulk import: it
requires `tenant.user_id IS NOT NULL` AND an accepted invitation,
both incompatible with landlord-only tenants (which have user_id
NULL and no invitation). Every bulk-imported lease in production
was being rejected with a misleading "Tenant not found" error.

Fix: drop the call entirely and replace with the actual invariant
the cycle-2 H9 finding intended — an inline overlap check that
rejects any new lease whose date range collides with an existing
active/draft/pending_signature lease on the same unit. Migration
20260423120000_v23_fourth_audit_lease_overlap.sql applied.

MEDIUM:
- M1: All three components were over the 300-line CLAUDE.md cap.
  - bulk-import-stepper.tsx (332 → 198) — extracted state to
    useBulkImportStepperState hook
  - bulk-import-result-panel.tsx (325 → 216) — extracted
    ResultErrorList + buildFailedRowsCsv to its own file
  - documents-section.tsx (310 → 240) — extracted validateFiles +
    reportUploadSummary helpers to documents-section-helpers.ts
- M2: handleComplete in bulk-import-dialog now memoized via
  useCallback so the auto-close 5-second timer in the stepper's
  useEffect isn't torn down on every parent re-render.

LOW:
- L1: buildFailedRowsCsv now falls through to the synthetic
  `row,error` shape when every error has row === 0 (mutation
  threw, no real CSV row to retry). Previously the
  parseResult-non-null branch emitted a header-only empty download.
- L2: Added NULL guard for p_payment_day in the RPC matching the
  pattern used for the other required params (`NULL < 1` evaluates
  to NULL → range check skipped → NOT NULL violation downstream).
- L3: Dropped redundant resetDialog() calls after onComplete().
  The stepper's internal state resets on remount via fresh useState
  initialisers; the previous code was dispatching setState on an
  about-to-unmount tree.
- L4: Cross-checked template-definition-keys.ts — false positive
  (path substring match on `documents/templates/` unrelated to the
  vault list shape change).

NIT:
- N1: Cumulative reset lifted to a `resetImportState` helper inside
  the new hook, removing duplication between handleFileSelect /
  handleBack-validate / handleBack-confirm.
- N2: Skipped — pre-existing convention (bare-tuple all() in
  ownerDashboardKeys) is consistent with the rest of the codebase
  per the cycle-4 cross-check section. No fix needed.

Tests:
- New test file bulk-import-result-error-list.test.ts pins down
  the L1 synthetic-error contract so a future refactor can't
  reintroduce the empty-CSV bug.
- 7,416 unit tests passing (+44 from cycle-3 baseline).

[hudsor01]
